requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==3.1
jdcal==1.4.1
lxml==4.6.2
macholib==1.14
openpyxl==3.0.6
Pillow==8.1.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
requests==2.25.1
selenium==3.141.0
soupsieve==2.2
urllib3==1.26.3
